fix basilisk_should_swap_assets_coming_from_karura_when_onchain_route_present integration test
