+++ integration-tests/src/oracle.rs
@@ -11,18 +11,6 @@ use hydradx_traits::{AggregatedPriceOracle, OraclePeriod::*};
 use pallet_ema_oracle::OracleError;
 use xcm_emulator::TestExt;
 
-pub fn basilisk_run_to_next_block() {
-	let b = basilisk_runtime::System::block_number();
-
-	basilisk_runtime::System::on_finalize(b);
-	basilisk_runtime::EmaOracle::on_finalize(b);
-
-	basilisk_runtime::System::on_initialize(b + 1);
-	basilisk_runtime::EmaOracle::on_initialize(b + 1);
-
-	basilisk_runtime::System::set_block_number(b + 1);
-}
-
 #[test]
 fn xyk_trades_are_ingested_into_oracle() {
 	TestNet::reset();
